--- Quoridor/src/graphics/GameController.java
@@ -1,5 +1,0 @@
-package graphics;
-
-public class GameController {
-
-}
