--- Quoridor/src/graphics/SettingsController.java
@@ -1,4 +1,0 @@
-package graphics;
-
-public class SettingsController {
-}
