+++ Quoridor/src/graphics/settings.fxml
@@ -8,7 +8,7 @@
 
 <AnchorPane xmlns="http://javafx.com/javafx"
             xmlns:fx="http://javafx.com/fxml"
-            fx:controller="graphics.SettingsController"
+            fx:controller="graphics.Controller"
             prefHeight="400.0" prefWidth="600.0">
 
 </AnchorPane>
